+++ test/resources/checkers/c_c++-clang-blocks.c
@@ -1,0 +1,8 @@
+#include <stdio.h>
+
+typedef void (^my_block_t)();
+
+int main(void)
+{
+    my_block_t p = ^ { printf("hello world\n"); };
+}